Split up international models and panels to make files smaller and easier to work with
