+++ great_international/translation.py
@@ -1,10 +1,15 @@
 from modeltranslation.decorators import register
 
 from core.translation import BaseTranslationOptions
-from great_international import models
+from great_international.models import (
+    great_international,
+    invest,
+    capital_invest,
+    find_a_supplier,
+)
 
 
-@register(models.InternationalSectorPage)
+@register(great_international.InternationalSectorPage)
 class InternationalSectorPageTranslationOptions(BaseTranslationOptions):
     fields = (
         'heading',
@@ -66,7 +71,7 @@ class InternationalSectorPageTranslationOptions(BaseTranslationOptions):
     )
 
 
-@register(models.InternationalArticlePage)
+@register(great_international.InternationalArticlePage)
 class InternationalArticlePageTranslationOptions(BaseTranslationOptions):
     fields = (
         'article_title',
@@ -80,7 +85,7 @@ class InternationalArticlePageTranslationOptions(BaseTranslationOptions):
     )
 
 
-@register(models.InternationalCampaignPage)
+@register(great_international.InternationalCampaignPage)
 class InternationalCampaignPageTranslationOptions(BaseTranslationOptions):
     fields = (
         'campaign_subheading',
@@ -117,8 +122,8 @@ class InternationalCampaignPageTranslationOptions(BaseTranslationOptions):
     )
 
 
-@register(models.InternationalHomePage)
-@register(models.InternationalHomePageOld)
+@register(great_international.InternationalHomePage)
+@register(great_international.InternationalHomePageOld)
 class InternationalHomePageTranslationOptions(BaseTranslationOptions):
     fields = (
         'hero_title',
@@ -174,7 +179,7 @@ class InternationalHomePageTranslationOptions(BaseTranslationOptions):
     )
 
 
-@register(models.InternationalArticleListingPage)
+@register(great_international.InternationalArticleListingPage)
 class InternationalArticleListingPage(BaseTranslationOptions):
     fields = (
         'landing_page_title',
@@ -184,12 +189,12 @@ class InternationalArticleListingPage(BaseTranslationOptions):
     )
 
 
-@register(models.InternationalRegionPage)
+@register(great_international.InternationalRegionPage)
 class InternationalRegionPageTranslationOptions(BaseTranslationOptions):
     fields = []
 
 
-@register(models.InternationalTopicLandingPage)
+@register(great_international.InternationalTopicLandingPage)
 class InternationalTopicLandingPageTranslationOptions(BaseTranslationOptions):
     fields = (
         'landing_page_title',
@@ -198,7 +203,7 @@ class InternationalTopicLandingPageTranslationOptions(BaseTranslationOptions):
     )
 
 
-@register(models.InternationalCuratedTopicLandingPage)
+@register(great_international.InternationalCuratedTopicLandingPage)
 class InternationalCuratedTopicLandingPageTranslationOptions(
     BaseTranslationOptions
 ):
@@ -225,7 +230,7 @@ class InternationalCuratedTopicLandingPageTranslationOptions(
     )
 
 
-@register(models.InternationalGuideLandingPage)
+@register(great_international.InternationalGuideLandingPage)
 class InternationalGuideLandingPageTranslationOptions(
     BaseTranslationOptions
 ):
@@ -249,26 +254,26 @@ class InternationalGuideLandingPageTranslationOptions(
     )
 
 
-@register(models.InternationalLocalisedFolderPage)
+@register(great_international.InternationalLocalisedFolderPage)
 class InternationalRegionalFolderPageTranslationOptions(
     BaseTranslationOptions
 ):
     fields = []
 
 
-@register(models.InternationalEUExitFormPage)
+@register(great_international.InternationalEUExitFormPage)
 class InternationalEUExitFormPageTranslationOptions(BaseTranslationOptions):
     fields = []
 
 
-@register(models.InternationalEUExitFormSuccessPage)
+@register(great_international.InternationalEUExitFormSuccessPage)
 class InternationalEUExitFormSuccessPageTranslationOptions(
     BaseTranslationOptions
 ):
     fields = []
 
 
-@register(models.InternationalCapitalInvestLandingPage)
+@register(capital_invest.InternationalCapitalInvestLandingPage)
 class InternationalCapitalInvestLandingPageTranslationOptions(
     BaseTranslationOptions
 ):
@@ -314,7 +319,7 @@ class InternationalCapitalInvestLandingPageTranslationOptions(
     )
 
 
-@register(models.CapitalInvestRegionPage)
+@register(capital_invest.CapitalInvestRegionPage)
 class CapitalInvestRegionPageTranslationOptions(
     BaseTranslationOptions
 ):
@@ -382,7 +387,7 @@ class CapitalInvestRegionPageTranslationOptions(
     )
 
 
-@register(models.CapitalInvestOpportunityListingPage)
+@register(capital_invest.CapitalInvestOpportunityListingPage)
 class CapitalInvestOpportunityListingPageTranslationOptions(
         BaseTranslationOptions):
     fields = (
@@ -391,7 +396,7 @@ class CapitalInvestOpportunityListingPageTranslationOptions(
     )
 
 
-@register(models.CapitalInvestOpportunityPage)
+@register(capital_invest.CapitalInvestOpportunityPage)
 class CapitalInvestOpportunityPageTranslationOptions(
     BaseTranslationOptions
 ):
@@ -452,7 +457,7 @@ class CapitalInvestOpportunityPageTranslationOptions(
     )
 
 
-@register(models.CapitalInvestRelatedRegions)
+@register(capital_invest.CapitalInvestRelatedRegions)
 class RelatedRegionsSerializer(
         BaseTranslationOptions):
     fields = (
@@ -460,7 +465,7 @@ class RelatedRegionsSerializer(
     )
 
 
-@register(models.CapitalInvestRelatedSectors)
+@register(capital_invest.CapitalInvestRelatedSectors)
 class RelatedSectorsSerializer(
         BaseTranslationOptions):
     fields = (
@@ -468,7 +473,7 @@ class RelatedSectorsSerializer(
     )
 
 
-@register(models.CapitalInvestRegionCardFieldsSummary)
+@register(capital_invest.CapitalInvestRegionCardFieldsSummary)
 class CapitalInvestRegionCardFieldSerializer(
         BaseTranslationOptions):
     fields = (
@@ -476,7 +481,7 @@ class CapitalInvestRegionCardFieldSerializer(
     )
 
 
-@register(models.CapitalInvestHomesInEnglandCardFieldsSummary)
+@register(capital_invest.CapitalInvestHomesInEnglandCardFieldsSummary)
 class CapitalInvestHomesInEnglandCardFieldsSummarySerializer(
         BaseTranslationOptions):
     fields = (
@@ -484,7 +489,7 @@ class CapitalInvestHomesInEnglandCardFieldsSummarySerializer(
     )
 
 
-@register(models.InvestInternationalHomePage)
+@register(invest.InvestInternationalHomePage)
 class InvestHomePageTranslation(BaseTranslationOptions):
     fields = (
         'breadcrumbs_label',
@@ -570,14 +575,14 @@ class InvestHomePageTranslation(BaseTranslationOptions):
     )
 
 
-@register(models.InvestHighPotentialOpportunityFormPage)
+@register(invest.InvestHighPotentialOpportunityFormPage)
 class HighPotentialOpportunityFormPageTranslationOptions(
     BaseTranslationOptions
 ):
     fields = []
 
 
-@register(models.InvestHighPotentialOpportunityDetailPage)
+@register(invest.InvestHighPotentialOpportunityDetailPage)
 class HighPotentialOpportunityDetailPageTranslationOptions(
     BaseTranslationOptions
 ):
@@ -635,14 +640,14 @@ class HighPotentialOpportunityDetailPageTranslationOptions(
     )
 
 
-@register(models.InvestHighPotentialOpportunityFormSuccessPage)
+@register(invest.InvestHighPotentialOpportunityFormSuccessPage)
 class HighPotentialOpportunityFormSuccessPageTranslationOptions(
     BaseTranslationOptions
 ):
     fields = []
 
 
-@register(models.InvestSectorPage)
+@register(invest.InvestSectorPage)
 class InvestSectorPageTranslationOptions(BaseTranslationOptions):
     fields = (
         'description',
@@ -681,14 +686,14 @@ class InvestSectorPageTranslationOptions(BaseTranslationOptions):
     )
 
 
-@register(models.InvestRegionLandingPage)
+@register(invest.InvestRegionLandingPage)
 class RegionLandingPageTranslationOptions(BaseTranslationOptions):
     fields = (
         'heading',
     )
 
 
-@register(models.InternationalTradeHomePage)
+@register(find_a_supplier.InternationalTradeHomePage)
 class InternationalTradeHomePageTranslationOptions(BaseTranslationOptions):
     fields = (
         'breadcrumbs_label',
@@ -712,7 +717,7 @@ class InternationalTradeHomePageTranslationOptions(BaseTranslationOptions):
     )
 
 
-@register(models.InternationalTradeIndustryContactPage)
+@register(find_a_supplier.InternationalTradeIndustryContactPage)
 class InternationalTradeIndustryContactPageTranslationOptions(
     BaseTranslationOptions
 ):
